refactor(view): board 판 분리
